+++ src/test/java/com/wounit/annotations/TestUnderTest.java
@@ -19,13 +19,13 @@ package com.wounit.annotations;
 import static org.hamcrest.CoreMatchers.hasItem;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
-import static org.junit.Assert.assertThat;
+import static org.hamcrest.MatcherAssert.assertThat;
 
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.extension.RegisterExtension;
 import org.mockito.Spy;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.webobjects.foundation.NSArray;
 import com.wounit.model.FooEntity;
@@ -37,9 +37,9 @@ import er.extensions.foundation.ERXArrayUtilities;
 /**
  * @author <a href="mailto:hprange@gmail.com.br">Henrique Prange</a>
  */
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class TestUnderTest {
-    @Rule
+    @RegisterExtension
     public final MockEditingContext mockEditingContext = new MockEditingContext("Test");
 
     @UnderTest(size = 2)